+++ src/gb/V8/gbMemory.cpp
@@ -347,7 +347,7 @@ mapperMBC3 gbDataMBC3 = {
 	0, // timer latched hours
 	0, // timer latched days
 	0, // timer latched control
-	time_t(-1) // last time
+	u32(time_t(-1)) // last time
 };
 
 void memoryUpdateMBC3Clock()
@@ -1334,7 +1334,7 @@ mapperTAMA5 gbDataTAMA5 = {
 	0, // timer latched months
 	0, // timer latched years
 	0, // timer latched control
-	time_t(-1) // last time
+	u32(time_t(-1)) // last time
 };
 
 void memoryUpdateTAMA5Clock()
